Add filter to canonicalize authors, affiliations
The filter turns metadata entries for authors and their affiliations
into a canonical form. This allows users to conveniently declare
document authors and their affiliations, while making it possible to
rely on default object metadata structures when using the data in other
filters or when accessing the data from custom templates.